font changes on figure
I also deleted the aliases to the datafiles.  There is a bug between
Dropbox and OS X where it gets stuck in an infinite loop trying to find
the files (causing a process to take up a huge chuck of memory), which
is impossible since they aren't actually stored locally.  Sorry for the
inconvenience!